Posts are now saved with verification information on creation
A post will be marked as verified if the password entered into the
new discussion/reply form matches the password in the database (uses
the application's encrypt and encrypt-verify utility functions).

When a post is shown in a thread, it will say either 'verified: 0' or
'verified: 1' for not verified and verified, respectively. This should
be made more user-friendly once the front-end receives some sort of design.

At the moment, doing something such as not entering a password in the form
is not handled, and will likely throw an error.